Concept of Static Initializer  Block
Static initialization block will execute at first without any object creation and that block will check the values of B and H are positive or not then according to the result it sets the value of flag to true/false. The main method receives falg's value and according to that value, it will calculate the area.